# docs/swagger.yaml
basePath: /
definitions:
...
  api.searchTicketsRequest:
    properties:
      departure_date:
        type: string
      destination_city_id:
        type: integer
      origin_city_id:
        type: integer
      vehicle_type:
        enum:
        - BUS
        - TRAIN
        - AIRPLANE
        type: string
    required:
    - departure_date
    - destination_city_id
    - origin_city_id
    - vehicle_type
    type: object
...
  db.CheckReservationTicketStatus:
    enum:
    - RESERVED
    - NOT_RESERVED
    type: string
    x-enum-varnames:
    - CheckReservationTicketStatusRESERVED
    - CheckReservationTicketStatusNOTRESERVED
...
  db.Ticket:
    properties:
      amount:
        type: integer
      arrival_time:
        type: string
      count_stand:
        type: integer
      created_at:
        type: string
      departure_time:
        type: string
      id:
        type: integer
      route_id:
        type: integer
      seat_id:
        type: integer
      status:
        $ref: '#/definitions/db.CheckReservationTicketStatus'
      vehicle_id:
        type: integer
      vehicle_type:
        $ref: '#/definitions/db.VehicleType'
    type: object
...
  db.VehicleType:
    enum:
    - BUS
    - TRAIN
    - AIRPLANE
    type: string
    x-enum-varnames:
    - VehicleTypeBUS
    - VehicleTypeTRAIN
    - VehicleTypeAIRPLANE
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for Bilitio ticket booking system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:3000/swagger/index.html
  title: Bilitio API
  version: "1.0"
paths:
...
  /tickets:
    get:
      description: Retrieve all tickets from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Ticket'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tickets
      tags:
      - tickets
  /tickets/{ticket_id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed info for a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ticket details
      tags:
      - tickets
  /tickets/search:
    post:
      consumes:
      - application/json
      description: Search tickets by origin, destination, date, and vehicle type
      parameters:
      - description: Search tickets request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.searchTicketsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Ticket'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search tickets
      tags:
      - tickets
  /tickets/user/completed:
    get:
      description: Get all tickets the authenticated user has completed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Ticket'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get completed tickets for user
      tags:
      - tickets
  /tickets/user/not-completed:
    get:
      description: Get all tickets the authenticated user has reserved but not completed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Ticket'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get not completed tickets for user
      tags:
      - tickets
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"